Simplify cached find API
