+++ tests/e2e/Adapter/Base.php
@@ -65,11 +65,6 @@ abstract class Base extends TestCase
      */
     abstract protected function deleteIndex(string $collection, string $index): bool;
 
-    protected function supportsCachedFind(): bool
-    {
-        return true;
-    }
-
     public function setUp(): void
     {
         if (is_null(self::$authorization)) {
