+++ tests/e2e/Adapter/RedisTest.php
@@ -114,11 +114,6 @@ class RedisTest extends Base
         return true;
     }
 
-    protected function supportsCachedFind(): bool
-    {
-        return false;
-    }
-
     /**
      * Inherited test exercises the case where an INTEGER column is altered
      * to VARCHAR. Redis stores documents as JSON; type changes do not
